platform/src/requirements.txt
fastapi~=0.115
uvicorn[standard]~=0.29
PyYAML~=6.0.1
pydantic~=2.7
pandas~=2.2
matplotlib~=3.8
reportlab~=4.2
psycopg2-binary~=2.9.9
python-dotenv~=1.0
requests~=2.32
gunicorn~=22.0
prometheus-client~=0.20
sqlalchemy~=2.0.23
alembic~=1.12.1
psutil~=5.9.6
rich~=13.7.0
pytz~=2023.3

# New dependencies for enhanced functionality
redis==4.6.0
hiredis==2.2.3
redis-py-cluster==2.1.3
cryptography~=41.0.0
PyJWT~=2.8.0
cerberus~=1.3.5
jsonschema~=4.19.0
python-multipart~=0.0.6
aiofiles~=23.2.1
httpx~=0.25.0
structlog~=23.1.0
tenacity~=8.2.3
click~=8.1.7
typer~=0.9.0

# AWS and Configuration Management
boto3~=1.34.0
botocore~=1.34.0

# Additional dependencies for integration guide
aioredis~=2.0.1
asyncpg~=0.30.0
passlib[argon2]~=1.7.4
python-jose~=3.5.0
slowapi~=0.1.9
opentelemetry-api~=1.36.0
opentelemetry-instrumentation-fastapi~=0.57b0

# Development and testing
pytest~=7.4.0
pytest-asyncio~=0.21.1
pytest-cov~=4.1.0
black~=23.7.0
isort~=5.12.0
flake8~=6.0.0
mypy~=1.5.1
pre-commit~=3.4.0
bandit~=1.7.5

# Documentation
mkdocs~=1.5.2
mkdocs-material~=9.2.6
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
